Make installed Xyce headers self-contained
One use case is for users of Xyce to link to Xyce libraries.  To date
all people who have needed to do this are folks who have built Xyce
from source code already.

We are now in a collaboration with some folks who are unable to build
Xyce from source and need us to provide precompiled,
shared-library-enabled builds of Xyce.

"make install" installs all the necessary Xyce headers and libaries to
accomplish this, but until today we didn't realize that some of those
Xyce headers include Trilinos headers, and someone who hasn't built
Xyce from source probably doesn't have those.

Turns out that the include of Teuchos_RCP.hpp in N_CIR_Xyce.h was
completely unnecessary, and that was fixed in a previous commit.

However, fixing *that* simply exposed a second problem, that
N_DEV_DeviceSupport.h required Epetra_Util.h.  This wasn't an example
of an unnecessary include --- the DeviceSupport class made use of an
Epetra_Util object for the sole purpose of providing random numbers
for the "gain scale" homotopy.

Xyce has its own random numbers package, and I thought to use it
instead of Epetra_Util in this case.  That was surprisingly difficult
for several reasons:

  - Turns out that a Trilinos header that was getting included
    indirectly as a result of the Epetra_Util.h include was getting us
    a "#undef VERSION" in just the right place to hide a real problem.
    Configure scribbles a define of "VERSION" into Xyce_config.h, and
    if left defined this breaks several device models derived from
    Verilog-A, all of which were defining a VERSION variable.

  - Replacing the Epetra_Util object with a RandomNumbers object was
    another matter.  Repeated and incorrect reseeding of the system
    random number generator was breaking some tests.

    There is now one place in the top-level Xyce code where -randseed
    is applied to cause RandomNumber objects to be allocated and
    seeded identically.

With this commit, the header files needed to build an external program
that accesses Xyce shared libraries with the GeneralExternal API no
longer try to pull in any Trilinos headers, and one can build using an
installation of shared-library-enabled Xyce without having a copy of
Trilinos at all.