More optimizations, eliminate allocations/copies
Like most of these small optimizations, this one won't move the needle
very much.  Most of my recent commits have been to minimize copies of
expression-related strings during the parsing and setup phase of Xyce.

Doing this has required that the expression objects store a bit more
information, so this is a tradeoff between speed and memory usage at the
moment.    I need to add some code that will delete this extra memory
usage in the new expression library once the setup phase is over.