Remove Sacado from testGenCoup test harness
This test harness, which is a demonstration program of the General
External Coupling device, implements only very, very trivial devices.

Out of laziness, I had written it to use Sacado so that I didn't have
to construct the dFdX and dQdX matrices by hand.  But this program
does such simple devices (resistors, capacitors, voltage-controlled
voltage-sources, and so forth) that it would have been trivial to
write it without Sacado to begin with.

And having it not depend on Sacado means that one could, if one
wanted, compile this program on a system that has only a binary
install of Xyce (with shared library support and all Xyce libraries and
Xyce headers necessary) and no copy of Trilinos.

Replacing the original Sacado version with a non-sacado version.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/56