Fix another Epetra_Util.h breakage
N_CIR_XyceCInterface.C broke after I removed  the include of
Epetra_Util.h from N_DEV_DeviceSupport.h.  It was missing a <cstring>
include.

This only impacts some systems, which is why I'm only just getting it
fixed.  On my other system this wasn't a problem.  On Linux with gcc it
was.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/56