Move most param handling from std::vector to maps
This commit is an interim cleanup commit.  Unfortunatley, it makes the
setup phase for new expression slightly slower for now, even though I've
been trying to make this phase faster.

In general, I maintain STL containers of operators such as voltOp, as
it makes certain operations easier.    If I want to update the voltages
used by an expression it is easier to loop over (or std::find) on a
std::vector of voltOp RCPs than it is to traverse the whole AST looking
for them.  So, as I do the parsing, in addition to setting up and
allocating the AST, I also push_back operators onto containers like this.

In some cases, I use vectors, and in others I use maps.   Which I use
depends on various issues.

For some types of operators I attempt to maintain singletons, and
for others I do not.  For parameters, I thought I was doing a good job
of this, but it turns out that I am not.    During the initial parse
phase, I was allocating a single paramOp class for each unique param
name.  But when I attach external operators, I can't do this, and I have
to live with multiple paramOp classes associated with the same name.

I made some mistakes in how I did this, which this commit corrects.
The correction was easiest to do if I switched from using a std::vector
to using a std::unordered_map.   So as of this commit, in nearly
all functions where I do param lookups now, I use the map.

There is one exception to this.  For function arguments, I have to
maintain the order, so for any operations related to function args, I
have to use a vector (at least for now; possibly I'll come up with a
workaround).    So, unfortunately, I'm currently maintaining both a map
and vector of paramOps.  This slows down the setup phase, I believe, but
it is more correct.

The other thing this commit does is to move more of the N_UTL_Expression
function operations into the newExpression class.  In a lot of my recent
commits, I moved most of the "getVariables" and "getFunctions" etc logic
completely into the newExpression class.  This commit does the same with
"get_type".

The original reason for moving things into newExpression was efficiency,
but another reason occurrred to me today, and that is unit tests.  My
unit testsing is applied to newExpression, but NOT to N_UTL_Expression.
So by moving these functions more into the newExpression class, they are
in a place where they can be directly tested by unit tests.  I plan to
work on that next.  Much of my recent development has been slow b/c it
frequentyly modified N_UTL_Expression, and I couldn't "unit test" that
class, meaning I had to 100% rely on the much slower Xyce regression
tests.