Cleanup and refactoring, also fixed maps problem
This commit removes some old code including:

(1) the addToVariousArrays function, which was not being used and is now
gone.  This was an experiment to see if it would result in faster
regeneration of the various STL containers of AST nodes.  It was not
faster, in fact was much slower.  So, it is gone so it need not be
maintained.

(2) the get_type function.  I previously moved most of this function
from the N_UTL_Expression wrapper class into the newExpression class.
While doing that, I noticed that it was the only function that used the
lead current map, and it occured to me that I could get rid of that
container if I got rid of this function.  Then I did some investigating.
The function was only being called from N_TOP_CktNode_Dev.C, and it was
being called for silly reasons.    The function that called it was the
CktNode_Dev::getDepSolnVars function.

The logic of the function was as follows:

(A) get the Bsrc-oriented solution varialbes from the device instance.
ie, call deviceInstance_->getDepSolnVars().  This returns a std::vector
of strings of all the expression variable names from the device.  These
variable names could be from one or more expressions, in theory, but
usually there is just a single expression if the device is a Bsrc.

(B) After this, then for each variable in the vector, loop over *all* the
dependent params of the device, and call get_type on the corresponding
expression.  So, if we have multiple dependent params, and thus have
multiple expressions, we pretend that we don't know which expression is
associated with which variable, and just call get_type on every
expression, for every variable.  Like I said before, most (all?) Bsrc's
just have one expression, so in practice this won't matter much, but it
is sill.

(C) Then, once an expression successfully determines the type of the
variable, the loop over params stops, and the node name and its type are
pushed back onto the dsVars container.

All of the above was silly.  The solution variables, that are pulled out
of the instance, were 100% understood at the time that they were first
set up in the device entity.  The entity pulled together vectors of
voltage nodes, current device names, etc. to create the master list of
variables.  So, there should be no need to figure this out again,
especially using various "find" calls.

So, updated the code to delete the get_type function, and I have
modified the CktNode_Dev::getDepSolnVars function so that it just
queries the device instances for a vector of types that has the same
size and order as the vector of variables..  Then, it just uses
those to do push_backs onto dsVars.  This is simpler, doesn't require
any searches, and allows me to delete a function we didn't need.

Finally one other thing is accomplished with this commit, and that is to
fix some of the map container usage in the new expression library.

I have been going back and forth between using std::vector and
std::unordered_map to hold different groups of RCPs to AST node
classes.    Currently, for any node that I am attempting to treat as a
singleton, I am using unordered_maps.  This doesn't appear to be faster
than just using std::vector.  In fact it is a little slower.  So I will
probably switch back.  But before doing so, I found a bug in how I'm
allocating these maps and this commit fixes it.  The bug didn't appear
to be causing problems, but it was still wrong and is now correct.

As of this commit the same 3 tests are still failing as before, one
that Pete is fixing, and two that I've chosen to blow off for now.
As for performance, I've slowed things down with my recent work,
although that did result in real problems being fixed.  For the LDRD
45nm GF PDK circuit that I've been using as a benchmark, my best setup
time with the new library has been around 53 seconds, although the
fastest build of Xyce that I had  usually wasn't quite that fast.
With these recent changes the timing of this setup phase is back
up into the mid-60 second range, consistently.    I think switching
from std::vector to std::unordered_map on params caused this slow
down, so I will probably be switching back.