New comments pertaining to "make_var" fucnction
I started to investigate today if the loop inside of the
"newExpression::getValuesFromGroup_ function is really necessary.  Most
parameters are simply attached as external dependencies, so it isn't
necessary to "set" them here.  But there are exceptions.  The paramOps
that have been tagged with isVar=true need to be updated in this way.
I discovered this empirically, but just running some of the large
circuits to see if this loop is ever executed.

The short answer is that it is sometimes executed, but only on paramOps
with isVar=true.  These are parameters that do not have an attached
external node, and have also been declared to be "Vars" by the
"make_var" function call.

This led me to dig into the "make_var" function, and to better
understand why it is called.   I have concluded that it should not be
needed, but getting rid of it will require a little refactoring.

I have actually wondered for a while what "isVar" even meant, as the
meaning was confusing to me.  Now I know, and based on this knowledge I
believe it should go.

The "make_var" function is a holdover from the old expression library.
The parser goes thru a somewhat convoluted logic when it resolves
parameters and functions in an expression.  For parameters, it does this
work in the "CircuitContext::resolveStrings" function.  The "strings"
are unresolved entities from an expression, obtained via the
"newExpression::getUnresolvedParams" function.

For each "string" (better called an unresolved param), the function does
the following:  (1) check if the string can be found in the container of
.params.  (2) if it fails to match it to a .param, then check the
container of .global_params.  (3) if it fails to match it to a
.global_param, then call "make_var" on it.

It turns out that steps (1) and (2) only check the list of RESOLVED
params and global params.  If the string in question matches the name of
an unresolved param or global_param, then it fails to find it, and has
to call make_var.

So, basically, if make_var is called, all that means is that the
parameter being resolved was unlucky.    The .param or .global_param it
could have matched with just hadn't been resolved yet.

For the old expression library, this made sense.   The old expression
library did everything with string substitutions.  If a parameter wasn't
resolved yet, then the string substitution wasn't possible because the
string wasn't known yet.

For the new expression library, doing this doesn't make sense.  To attach one
expression to another, it doesn't matter if it (the attaching
expression) is resolved or not.  As long as it has been parsed and has
an AST tree, the node can be attached.  The attaching expression can
complete its resolution later.  So, I suspect that things will go
fine if I make the parser also attach "unresolved" params.

Doing that may screw up the logic in the parser.  I'll have to check
that, of course.