Document known defect of Gitlab-ex issue 60
This issue effects only rare use cases of Verilog-A analog function
use (using an analog function that has "side effect" output arguments,
then using these output arguments in such a way that their derivatives
are required, but not using the return value of the function the same
way).

When tripped, Xyce/ADMS will fail to compute the derivatives of the
output arguments at all, so they are always zero.

The issue is difficult to fix and will not be fixed or even worked on
before Xyce release 7.2, so I'm documenting it as a known defect in
the release notes.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/60