Change to table breakpoints, to be faster
This is related to bugs 1173 and 1176, which were about poor Xyce
performance, when handling sources using very large tables.  1173 was
about the PWL source, and the fix was primarily about refactoring how
Xyce handles breakpoints.

1176 is about the old expression library.  In the old expression
library, the breakpointing was not a bottleneck, but the setup of the
expression was a huge problem.  (the notorious PTcheck function was
involved again)

In my work on the new expression library, I was pretty aware of these
issues and I had partially addressed them.  The setup times for new
expression tables, for example, are very, very fast.

But as it turns out, for very large tables, my strategy for breakpoints
was flawed.  I had (mistakenly) thought that the best course of action
was to have the expression return ALL the table breakpoints in one
function call.  So, even if the table was 20,000 breakpoints, it would
load them all into the breakpoint array.  The good news is that since
this was only done 1x, there were no redundant breakpoints to process on
future time steps.  The bad news is that the breakpoints container was
now very large, so if the time integrator had to process it at all, it
was expensive.

Yesterday, Ting sent me a small customer circuit that used a very large
table in a Bsrc.  With the old expression library, on my laptop in
serial, this calculation had the following timings:

***** Problem read in and set up time: 143.128 seconds
***** Total Simulation Solvers Run Time: 4.37963 seconds
***** Total Elapsed Run Time:            147.507 seconds

Then, with the new expression library (using the all-in-one-go approach
to breakpoints) I got these timings:

***** Problem read in and set up time: 0.116176 seconds
***** Total Simulation Solvers Run Time: 28.6423 seconds
***** Total Elapsed Run Time:            28.7581 seconds

So, much better setup time (0.1 sec vs. 143 sec), but slower solver time
(28 sec vs. 4 sec).

This morning I refactored the breakpoint handling in the tableOp class.
Now, it no longer tries to do an all-in-one-go appraoch.  Instead, it
returns the nearest 2 breakpoints every time.  If they are redundant, so
be it; the time integrator can sort that out.  With these changes, I now
get the following timings for this circuit:

***** Problem read in and set up time: 0.093636 seconds
***** Total Simulation Solvers Run Time: 5.58202 seconds
***** Total Elapsed Run Time:            5.67558 seconds

So, the setup is still WAY faster, and the solver time is now in the
same ballpark.  I am not sure why the solver time is still slower
(although not by much), but that is an issue for another day.