Ignore IN(*) (for integer N) on .PRINT lines
This commit makes Xyce ignore (with a warning to stdout) any wildcard
request of the form IN(*), where N is an integer, on .PRINT lines.
This mostly affects the T device in Xyce.

This commit also excludes any of the power grid devices and the O
device from any I-, P-, or W-based wildcards.  Those devices do
not support lead currents or power calculations.

SOB Bugs 983 and 1317
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=983
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1317