Fix bison 3.2 dependence
The new expression library was originally written with a version of
Bison that is not available on our RHEL7 machine.  That version
produces an empty "position.hh" file for backward compatibility, but
the new expression library doesn't include that file and instead has a
"%require 3.2" restriction.

If one includes position.hh, which *IS* created by later versions of
bison but is empty, one can remove the %require "3.2" line.