Attempt to fold new expression into CMake build
The rebase of new-expression-integration needed to discard some work
that had been done on that branch for CMake integration, due to many
conflicts introduced by that work and simultaneous development on the
main develop branch.

This commit attempts to fold the small number of real changes needed
to make new expression work in the CMake build.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/38
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1249