Add missing file to CMakeLists.txt
Makefile.am had "ast.C" mixed in with the .h files later in the
SOURCES list.  Makefile.am requires all .C and .h files to be listed
here in order for "make dist" to work (this target supposedly builds a
complete tarball of the distribution, a feature we NEVER use).  CMake
doesn't have an equivalent feature and only lists the .C files.  Thus,
when looking over Makefile.am for the purposes of CMakeLists
reconciliation, I have a tendency to ignore things once the .h files
start showing up.

I overlooked ast.C when comparing Makefile.am and CMakeLists.txt,
because the .C didn't stand out in the field of .h files.

I have therefore added this file to CMakeLists.txt and moved it in
Makefile.am so that it is in among its other .C friends instead of
seated at the kids table.

I believe this will take care of the remaining missing symbols on
Windows.  Can only tell by pushing this change.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/38
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1249