Fix Segfaults in .PRINT Wildcard Handling
This commit fixes segfaults in the handling of .PRINT line wildcards
that use "multi-character operators" such as VR(1*),  VDB(1*), IR(V*)
and IDB(V*).  Those types of operators are now silently ignored.
This is intended as a short-term fix for the Xyce 7.2 release.
"Single character operators", such as V(1*) and I(V*), are properly
supported though.

SON Bug 1211
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1211