Partial Fix for Numerical Tol. for TD, FROM and TO
This commit provides a partial (and possibly temporary) fix for
the numerical roundoff issues that can occur when the FROM, TO or
TD qualifiers are expressions.  This fix only applies to TRAN
measure mode.  It does not fix this issue for AC, DC or NOISE
measure modes.  This quick fix is being done in support of SON
bug 1249, so that the new expression library branch can be merged
with the develop branch this week.

SON Bugs 1249 and 1318
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1249
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1318