Fix to allow "J" to be a parameter name
This was driven by the BUG_1784/adc_driver weekly test.  This circuit
has instances of 'M=J' throughout.

I had set aside "J" to mean the imaginary part of a complex number.  ie,
(1.0+2.0J), where the real part is 1 and the imaginary  part is 2.

But I did it in a restrictive way, which made it impossible for the
parser to handle "J" by itself, not preceded by a number.

This commit fixes that issue.  Now, a number immediately followed by "J"
is tokenized as TOK_IMAG_NUM, whereas other numbers not followed by "J"
are tokenized as TOK_NUM.  seems to work, at least in unit testing.

Also, part of what made this easier was my realization that a lot of
patterns that had the same action could be grouped together with the
or "|" symbol.  That makes the file a bit smaller.  I think I will do
more of this in a future commit.