Compressing more rules that have same actions
I recently (when fixing the issue with parameters named "J") realized that I
had a lot of groups of rules/regex patterns in the flex input file that resulted
in identical actions.  It is better to compress these into a single action
with a bunch of patterns OR'd together.  In other words, instead of
this:

pattern1 { action1 }
pattern2 { action1 }
pattern3 { action1 }

(where all 3 patterns have identical actions) do this:

pattern1 |
pattern2 |
pattern3 { action1 }

The above example is silly; in the real *.l file the "action1" code
goes for several lines, so only having it appear 1x is better and
cleaner.