Fix P(*1) and W(*1) for J, M, Q and Z Devices
This commit fixes that handling of .PRINT line wildcard syntaxes
of the form P(*1) and W(*1) for multiterminal devices like J, M
Q and Z.  The structure used to hold the device list, for each such
wildcard, must be a list rather than a vector.  Otherwise, the P()
or W() wildcard is duplicated; once for each lead current that the
device has.  A similar fix was done previously for the P(*) and
W(*) wildcards.

SON Bug 1211
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1211