Fix so that the TOK_WORD token can include periods
This is to address issue #66 in gitlab.

The TOK_WORD token is used fairly ubiquitously in the lexer, but it most
commonly means "parameter".    It relies on the "ID" macro to define its
regex pattern.  The ID macro did not, until this commit, include the
period or "dot" special character.

This commit adds the perdiod to ID, which makes it possible for
something like "eric.dat" to be a parameter name.

I also went ahead and added the period to the "DEV" macro, which is used
for device instances names.  The DEV macro is a bit more expansive
than ID in that it included all the characters of ID, plus a few special
characters like "+", which couldn't be used in ID as it would get
confused with the plus operator.

The dot "." was already in the NODE macro, which is the most expansive
macro in the lexer, and is used for voltage node names.  So, that macro
is unchanged and all the other macros are unchanged.

This commit also changes the section at the bottom of the file that
reports "unrecognized characters".  For some reason (probably before I
went back and fixed most of the error reporting), this error was simply
being printed via std::cout rather than invoking the Xyce error handler.
That is fixed now as well.

Finally, the parserUnitTest program now includes a unit test for
parmaeter names that include a period.