Fix failsafe for Nan to 1.0e+50 (it was 0.0)
This is to address issue #63

In issue #63, there was a test failure that involved the handling of
divide by zeros in the expression library.    The netlists involved had
resistors whose values were set via expressions, and the expressions had
(at the beginning of the simulation, and probably later) a divide by
zero problem.

In the new expression library I had created some failsafe fixes in the
evaluateFunction function.  These failsafes would do the following:

  result = astNodePtr_->val();
  Util::fixNan(result);
  Util::fixInf(result);

The first one, which fixed Nans would change result to 0.0.

The second one, which fixed Infs would change the result to 1.0e+50.

For the failing circuits, on most platforms, the divide by zero
expressions would return an Inf, and trigger the 1.0e+50.   On these
platforms the circuits would pass.

For some reason, on the failing platforms, the 1.0e+50 was not getting
applied.   Possibly, it was detecting a nan and turning it into zero
first.  Or, possibly something else was happening.  Anyway, the result
was (eventually) that result=0.0, and this caused singular matrices and
the simulation to fail.

After some discussion with Tom, he asked if setting a Nan to zero was
the appropriate fix.  We also discussed if the library should be
modified (in places like the binaryDivOp class) to avoid performing any
arithmetic that would produce Nan or Inf to begin with.

For now, I am experimenting with simply having Nans be set to 1.0e+50,
just like Infs are.  This appears to fix the failing tests.