Add "$" special character to TOK_WORD (parameters)
This is related to bug #66.

In a previous commit for that bug, I added the period special character
to the ID macro, which defines the regex used for the TOK_WORD token.
TOK_WORD is generally going to be parameter names (or basically any
string that isn't otherwise tokenized)

When I added the period, I also modified the function that reports any
"unrecognized symbol" errors.    Before, it had just gone to std::cout,
which was a mistake.  Now it (ultimately) goes to Xyce::Report, which
probably made it more strict.

After I pushed the "period" change, some new tests started failing.
Those tests were:

Certification_Tests/BUG_456/bug456pp
Certification_Tests/BUG_456/bug456pp_unpacked
SandiaTests/Certification_Tests/BUG_336/BUG_336
SandiaTests/C_T/BUG_SON_105/push_pull_xyce
TIA/TRAP/push_pullLte2

They all seemed to be variants of the same test.  The problem was that a
node name included the "$" symbol.    And, it was a node name that the
Xyce netlist parser had decided to parse via the expression library.

Note, one use case for the new expression library that I didn't
originally plan for, but that I eventually had to adapt to, was that
frequently the expression library is used to parse something.  And, ONLY
to parse something.  Once the parse is complete, then the resulting
expression isn't evaluated.  It is just to help out the netlist parser.

Usually, the expression library is used to help parse parameters, even
if the parameter doesn't utlimately wind up being expression-valued.  I
didn't think that it was used to parse raw node names, but apparently
this is sometimes done.  All the tests that failed were mutual inductor
test, so possibly some of the extra special parsing for mutual inductors
requires this.

Anyway, the new expression library has handled special, weird characters
in voltage nodes for a long time.  But, for those special characters to
be parsed properly, it is best if the lexer is able to tell from context
that these are voltage nodes and not parameters.  So, as the code was
written prior to this commit, {V($N_0077)} would parse fine, but
{$N_0077} would not.  The first case has "V(" before the node name, so
the lexer can see this and change context.  In the second case, it
doesn't have this information, so it stays in the main context and
parses it as a parameter.

The macro (regex) that I use for parameters is more restrictive than the
one that I use for voltage nodes.  This is b/c voltage nodes need to
support things like "+", which are also mathematical operators, and thus
a problem for parmaeter names.

But, fortunately, the "$" symbol isn't overloaded in this way.  So there
was no danger in adding it to the macro(regex) ID, so that it can be
parsed as a TOK_WORK (parameter) token.

So, this commit makes that change.  Now that I've made this change, on
my local machine all 2251 tests pass when I use these tags:

+nightly+serial?klu?rad?qaspr?fft