Fix "toy" diode models' diode level
These "toy" diode models written in Verilog-A are part of a sort of
user test suite for trying out Xyce/ADMS, and figure prominently in
the Xyce/ADMS users' guide.

Until now, these models tried to slot themselves into Xyce as the
level 2001 and 2002 diode.

Ever since I added Diode_CMC to the code, however, level 2002 is
already taken and trying to run with a plugin build according to
package directions would break.

This commit changes the diode level of the toys to something truly
enormous and unlikely to conflict with anything in the near future.

This was exposed only because I tried to test new-expression with
Xyce/ADMS toys (because the toy vsrc.va model allows time-dependent
expressions in its argument, and the DiodeClipper.cir netlist actually
tries to use that feature).

The good news is that once I fixed the breakage with model levels,
everything just worked as expected.

Not really related to
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/38 but found in the
process of last-minute testing thereof.