Fix Handling of .PRINT Wilcards like V(*1)
This commit fixes that handling of .PRINT line wildcards, like V(*1),
that end with a specific string.  When the wildcard ends with a
specific string, rather than a * character, then there must be an
exact match to that specific string at the end of the name string
being tested for a match.

SON Bug 1211
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1211