Removing old expression library files
Now that the new expression library has been  merged into Xyce, the old
expression library files are not needed anymore, and should be removed.
This commit removes them.  It also updates the relevant CMake and
Makefile.am (for autoconf) files.

The netlist parser had one small place in which the old expression
library header was included that I had previously overlooked.  It also
modified a tiny bit of (now unused, and unnecessary) static data in the old library.
That is now also removed in this commit.