Removing #include<newExpression.h>
This header, newExpression.h, was not needed in the
mainXyceExpressionGroup.h header.  It was causing problems for the
Charon build, indirectly.

mainXyceExpressionGroup.h is included into the N_CIR_Xyce.h file, which
means it is pulled into the Charon build.

newExpression.h is a problem, because it includes ast.h.

ast.h is a problem because it includes N_UTL_Interpolators.h.

N_UTL_Interpolators.h is a problem because it includes
Sacado_No_Kokkos.hpp, which conflicts with the version of Sacado used by
Charon.

But fortunately, newExpression.h doesn't need to be there.  So,
hopefully this commit will fix the problem.