Remove Teuchos and expr group from N_CIR_Xyce.h
This work was originally prompted by a broken Charon build earlier
today.  My last several commits fixed that issue, which was mostly
about conflicting Kokkos.

My fix removed that conflict, but didn't completely remove
trilinos dependencies.  Teuchos still needed to be in the
N_CIR_Xyce.h header.

Tom reminded me that it is important to keep Trilinos dependencies
out of this file, and that he had recently spent a lot of time
removing them, in support of an external customer who needed to use
Xyce as a library but couldn't build it.

Unfortunately, the merge of new-expression into develop a few days ago
broke this rule.   And, as this isn't tested directly nobody noticed.

This commit 100% removes trilinos dependencies from N_CIR_Xyce.h.
It turns out that the one thing in the file that used RCP, which was
the reason for the Teuchos header, didn't need to be there.  It only
needed to be in a single function inside N_CIR_Xyce.C.   So, this made
it pretty easy to fix.