Get rid of a bison deprecated warning
This commit is in partial support of #53, which is about updating the
Bison input file to use Bison variants.  Variants were first supported
in Bison in version 3.0.4.  We recently made the decision to have 3.0.4
be the minimum version for Xyce compilation.  It had previously been
2.4.

I have been using bison 3.5 on my local machine and it has been giving
me "deprecated" warnings about 2 structures in my ExpressionParser.yxx
file.  The warnings also recommend how to make them stop.   The
recommendations are:

go from:
%define parser_class_name "ExpressionParser"
to:
%define api.parser.class {ExpressionParser}

and also from:
%name-prefix="XyceExpression"
to:
%define api.prefix {XyceExpression}

I was going to change both of these, but I double checked the bison
documentation and discovered the following:

%define api.parser.class  was introduced in Bison 3.3

%define api.prefix was introduced in Bison 2.6

As noted above, we had previously been trying to maintain compatibility
with older versions of bison, back as far as 2.4.  That was to support
RHEL6, which had 2.4.  But, RHEL6 is no longer one of our supported
platforms, so now, the oldest version of Bison we need to support
is 3.0.4.

Based on that, I can and should change "prefix" but cannot change the
parser class definition.  So, this commit changes prefix and leaves the
parser class alone.