Remove kludge that was only necessary for 2.3-2.x?
Bug #53, indirectly.

Near the top of the bison input file, there was a block of code that I
had copied over (and modified) from the reaction parser.  The text is
below.

I never really understood very well what this was doing, but left it in
there b/c of the the comments.

Bug #53 is about using bison variants, which started being
supported in 3.0.4.    We recently decided to have 3.0.4 be the minimum
version of Bison needed to compile Xyce.  It was previously 2.4, as that
was the default version on RHEL6.

Now that 3.0.4 is the official minimum, it seemed to me that this
confusing block of code could go, so this commit removes it.  I have
checked and the expression library still compiles fine.

----- code block ------
  // WARNING!!!!  The stuff in this block is the "post-prologue" that
  // in versions of Bison prior to 2.3 used to be placed into the .hxx
  // file generated by bison, and which we could therefore just include
  // to get the right stuff.
  // In version 2.3 they stopped doing that and started putting the
  // post-prologue code directly into the generated .cxx file, which means
  // that we need to DUPLICATE this block of code in N_DEV_ReactionNetwork.C.
  // If you change *ANYTHING* in this post-prologue, you MUST edit
  // N_DEV_ReactionNetwork.C and keep it in sync with the changes in this
  // block!
  // Note also that this was absolutely critical to do in Bison 2.3, and
  // somehow didn't seem to matter much just a few releases of bison later.
  // Careless hacking of this block *could* break compatibility of Xyce with
  // older versions of Bison and go unnoticed in testing.  Please be careful.
  /* CAREFUL watch continuations! */
int ExpressionLexer::getToken(ExpressionParser::semantic_type *lvalp,  \
                            location *llocp)