Removes unnecessary linear algebra object headers
While trying to sort out the amount of work needed to help with
issue #76, I found several unnecessary inclusions of linear algebra
headers.  In particular, N_LAS_MultiVector, which is really not used
in the device package at all.  At the very least, these header
inclusions can be removed to eliminate confusion in the dependencies
of the device package to the linear algebra package.

Gitlab issue #76