Migrate .PRINT Wildcards (* and ?) to Use Functions in <regex>
This commit replaces the functions in N_IO_WildcardSupport with
ones that leverage functions defined in the standard <regex> header
file.  This change should make the code more maintainable and easier
to understand.  It also allows the support of the HSPICE-like ?
wildcard, which is defined to mean "any one character".  This change
does require C++11 support though.

This commit also makes, for example, V(1?) to be solely a wildcard
request for all two-character node names that start with '1'.  It
will NOT provide an explicit match for only the voltage at node 1?,
if node 1? exists in the netlist.  Finally, the ? character is not
considered to be a wildcard if V(1?) is used within an expression
on a .PRINT line.

SON Bugs 1211 and 1319
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1211
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1319