Change from character literal tokens to symbolic
Issue #53

This commit also contains several code re-arrangements.

The main thing in it is the removal of character literals as tokens,
such as '+'.   This was done to make %define api.token.raw possible,
which is a feature of Bison that became available in version 3.5.
This means that it cannot be used by default right now, in Xyce, which has
a minimum of 3.0.4.  So, although there is a "%define api.token.raw"
present in the *yxx file, it is commented out.

This commit also moves all of the %define statements to the top of the
file, away from the "%union" declaration, and away from the code blocks.
That seems to be (based on bison examples I've been seeing online) a
more standard way to organize a Bison file.

This commit also replaces the %{ %} way of specifying code blocks with
%code instead.  Using %code makes it easier to remove the %union
declaration, which will be a necessary aspect of using variants.