Got *yxx file to compile with variants
Issue #53

This commit represents the first time that the bison input file,
ExpressionParser.yxx, compiles with variants turned on.

WARNING: IT STILL DOESN'T WORK!!!   Among other things it is
inconsistent with the *l (Flex input) file, so ExpressionLexer.l will
not compile right now.

I had a lot of difficulty just getting this (the *yxx file) to
compile properly, so I wanted to do a checkpoint commit.

Part of the difficult was that when the %union declaration is taken
away, some of the %{ %} code literals are either re-arranged, or ignored
by Bison.  So, that caused some problems.  I fixed some of the problem
by using %code instead, which allows you to be more specific, and makes
it so Bison doesn't ignore it.

Another difficulty was that I initially set both of these, because I saw
them in numerous examples:

%define api.value.type variant
%define api.token.constructor

I did not understand the second one (still don't, for the most part) and
it caused the yylex function  (renamed in this file as
XyceExpressionlex) to only be called with a single argument from
ExpressionParser.yxx instead of 3, which is what I had been doing
before.   So, now the api.token.constructor define is commented out.  I
hope to figure it out later.  If it is beneficial, I'll use it, but for
now I don't get it.

Another note:  the ultimate point of switching to variants is so that I
can get rid of the pointers-to-RCP objects, that I currently have to use
b/c of the %union.  I want to just use pure RCPs.  However, for the
purposes of this first "variant" change, I am still using
pointers-to-RCPs.  Assuming that they work OK, for my next step I plan
to fix the flex input file, ExpressionLexer.l to use the variant form.
I will switch from pointers-to-RCPs to pure RCPs only after I have that
working (assuming that is even possible).